feat: ui-graph-editor-orchestration-split-004 - Current-activity card as thin viewport
