+++ docs/internal/processes/factory-workstation-relevant-files.md
@@ -118,6 +118,7 @@ work inputs.
 - Graph-editor session orchestration (edit mode, active tool, classifier gating, `canInteractWithEditor`, mode toggle) lives in `ui/src/features/workflow-activity/hooks/use-graph-editor-session.ts`; `useCurrentActivityGraphEditor` keeps `editorMode`/`activeTool` state for `useCurrentActivityEditableGraph({ editorMode })` and composes the session hook with a `leaveEditorRef` callback so clean leave resets transient controller state without circular hook ordering.
 - Graph-editor save/leave orchestration (confirm-save/leave flags, active-work and stale-draft guards, save/discard/leave handlers, transient reset) lives in `ui/src/features/workflow-activity/hooks/use-graph-editor-save-flow.ts` using `buildFactoryGraphSaveSummary` and `getFactoryGraphEditorMessages`; compose it after session/controllers with `leaveEditorRef` and `setIsConfirmingLeaveEditorRef` so mode-toggle leave confirmation does not circularly depend on save-flow state; prove guards and discard-without-exit in `use-graph-editor-save-flow.test.ts`.
 - Graph-editor tool controllers (connect, remove, add entity) live in `ui/src/features/workflow-activity/hooks/use-graph-editor-controllers.ts`, composing `react-flow-current-activity-card-editor-connections.ts`, `react-flow-current-activity-card-editor-removals.ts`, and `useFactoryGraphAddEntityController` from `react-flow-current-activity-card-editor-chrome.tsx` without changing `factory-graph-editor` operation semantics; prove connect/removal behavior in `react-flow-current-activity-card-editor-controllers.test.tsx` and composition in `use-graph-editor-controllers.test.ts`.
+- Current-activity graph card composition stays thin in `ui/src/features/workflow-activity/components/react-flow-current-activity-card.tsx` (session + editor hook, import/save/dialog chrome wiring only); React Flow viewport wiring lives in `react-flow-current-activity-card-viewport.tsx`; observer/editor graph projection and node/edge state live in `hooks/react-flow-current-activity-card-graph-view-model.ts` with shared selection types in `lib/react-flow-current-activity-card-types.ts`; prove loading/empty/error, accessibility, and edit→save/discard/leave flows in `react-flow-current-activity-card.test.tsx`, `react-flow-current-activity-card-surface.test.tsx`, and `react-flow-current-activity-card-edit-integration.test.tsx`.
 - Workflow-activity tests that render localized shells such as `workflow-activity-bento-card.tsx` should derive expected accessible names from the same message catalog helpers the component uses, rather than hardcoding English viewport or toolbar labels when the test passes a non-default `locale`.
 - Graph-editor removal browser verification should stay split across focused Storybook seams: destructive confirmation and blocked-removal notice states belong in `ui/src/features/factory-graph-editor/components/factory-graph-editor-controls.stories.tsx`, while pending-removal node and edge rendering belongs in `ui/src/features/factory-graph-editor/components/factory-graph-editor-flow.stories.tsx`.
 - Graph-editor add-entity browser verification should stay in focused dialog seams: prove work-type creation, work-state creation, and field-level validation copy in `ui/src/features/factory-graph-editor/components/factory-graph-editor-add-dialog.stories.tsx`, while keeping draft-mutation orchestration in `ui/src/features/workflow-activity/components/react-flow-current-activity-card.test.tsx` instead of widening dashboard-wide browser suites.
