+++ ui/src/features/workflow-activity/lib/react-flow-current-activity-card-types.ts
@@ -1,0 +1,5 @@
+export type CurrentActivitySelection =
+  | { kind: "node"; nodeId: string }
+  | { kind: "state-node"; placeId: string }
+  | { kind: "worker"; workerName: string }
+  | { kind: "work-item"; dispatchId: string; nodeId: string; workID: string };
